+++ Web/Scotty/Action.hs
@@ -583,7 +583,7 @@ instance Parsable () where
 instance (Parsable a) => Parsable [a] where parseParam = parseParamList
 
 instance Parsable Bool where
-    parseParam t = if t' == TL.toCaseFold "true"
+    parseParam t = if t' == TL.toCaseFold "true" || t' == TL.toCaseFold "on"
                    then Right True
                    else if t' == TL.toCaseFold "false"
                         then Right False