+++ changelog.md
@@ -6,6 +6,7 @@
 * Add tests for URL encoding of query parameters and form parameters. Add `formData` action for decoding `FromForm` instances (#321).
 * Add explicit redirect functions for all redirect status codes.
 * Added ActionM variants for cookie module functions (#406).
+* Updated `Parsable` instance for `Bool` to accept `on` for HTML form checkboxes.
 
 ### Breaking changes
 * Remove dependency on data-default class (#386). We have been exporting constants for default config values since 0.20, and this dependency was simply unnecessary.